requirements.txt
random
sys
Figlet
requests
json
pytest